backend/requirements.txt
mcp==1.26.0
fastapi>=0.110.0
uvicorn
pydantic>=2.11.0
python-dotenv
openai
pinecone-client
requests
numpy
sentence-transformers